Updated Plugins (Users, Emails, Collections). Added Swagger Documentation for Plugins (Users, Emails, Collections)
