+++ src/Pluggins/collections/dto/addJsonSectionsResponse.dto.ts
