+++ src/Pluggins/collections/dto/deleteJsonSectionsResponse.dto.ts
