+++ src/Pluggins/collections/dto/getCompleteJsonByIdResponse.dto.ts
